+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2 (unreleased)
 
 - Added `summary` method
+- Added `parallel` option to `train` method
 
 ## 0.1.1 (2020-10-10)